# src/server/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"FileDownloadDetails": {
				"properties": {
					"name": {
						"type": "string"
					},
					"extension": {
						"type": "string"
					},
					"fullPath": {
						"type": "string"
					},
					"filePath": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"extension",
					"fullPath",
					"filePath"
				],
				"type": "object"
			},
			"ApiResponse_FileDownloadDetails_": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string"
					},
					"result": {
						"$ref": "#/components/schemas/FileDownloadDetails"
					}
				},
				"required": [
					"code",
					"type",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FileType": {
				"enum": [
					"audio",
					"video"
				],
				"type": "string"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "vytc",
		"version": "1.0.0",
		"description": "Converts youtube videos or other MP4 videos  MP3 format.",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "UcefBen",
			"url": "https://<git-repository-url>"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/downloader": {
			"get": {
				"operationId": "GetDownloadVideo",
				"responses": {
					"200": {
						"description": "Returns the file to be downloaded.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FileDownloadDetails_"
								}
							}
						}
					}
				},
				"description": "Download the the requested file with the desired format and name.",
				"tags": [
					"downloader"
				],
				"security": [],
				"parameters": [
					{
						"description": "The video key returned by the fetch action.",
						"in": "query",
						"name": "key",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The title of the video to be downloaded with.",
						"in": "query",
						"name": "title",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The type of the file [Audio || Video].",
						"in": "query",
						"name": "type",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/FileType"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}